requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database connectivity
sqlalchemy>=2.0.0
pyodbc>=4.0.39              # Voor SQL Server (primary database)

# Configuration & Logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Jupyter & Analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
